Add a lock screen
feat: add lock screen and shared useCurrentTime hook
