+++ src/App.jsx
@@ -25,6 +25,7 @@ import { SettingsProvider } from "./contexts/SettingsContext";
 import { ConnectorProvider } from "./contexts/ConnectorContext";
 import React from "react";
 import PairingScreen from "./components/auth/PairingScreen";
+import LockView from "./components/lock/LockView";
 
 export const NetworkContext = React.createContext({
   selectedNetwork: null,
@@ -447,6 +448,11 @@ function App() {
       if (albumImage) {
         updateGradientColors(albumImage, "nowPlaying");
       }
+    } else if (activeSection === "lock") {
+      const albumImage = currentlyPlayingAlbum?.images?.[1]?.url;
+      if (albumImage) {
+        updateGradientColors(albumImage, "lock");
+      }
     }
   }, [
     activeSection,
@@ -603,6 +609,8 @@ function App() {
         onNavigateToAlbum={handleNavigateToAlbumFromNowPlaying}
       />
     );
+  } else if (activeSection === "lock") {
+    content = (<LockView onClose={() => setActiveSection("recents")}/>)
   } else if (viewingContent) {
     content = (
       <ContentView
