+++ src/components/common/navigation/Sidebar.jsx
@@ -5,6 +5,7 @@ import {
   ArtistsIcon,
   RadioIcon,
   SettingsIcon,
+  LockIcon,
 } from "../../common/icons";
 import StatusBar from "./StatusBar";
 import { useSettings } from "../../../contexts/SettingsContext";
@@ -29,7 +30,7 @@ export default function Sidebar({ activeSection, setActiveSection }) {
       )}
       <div className="mr-4 flex-shrink-0">
         <div className="h-[70px] w-[70px] bg-white/25 rounded-[12px] flex items-center justify-center border border-white/10 drop-shadow-[0_20px_5px_rgba(0,0,0,0.25)]">
-          <Icon className="h-10 w-10" />
+          <Icon className="h-10 w-10 text-white" />
         </div>
       </div>
       <div>
@@ -54,6 +55,7 @@ export default function Sidebar({ activeSection, setActiveSection }) {
       <SidebarItem section="artists" icon={ArtistsIcon} label="Artists" />
       <SidebarItem section="radio" icon={RadioIcon} label="Radio" />
       <SidebarItem section="settings" icon={SettingsIcon} label="Settings" />
+      <SidebarItem section="lock" icon={LockIcon} label="Lock"/>
     </div>
   );
 }
